+++ mace/calculators/__init__.py
@@ -1,7 +1,7 @@
 from .foundations_models import mace_anicc, mace_mp, mace_off, mace_omol, mace_polar
 from .lammps_mace import LAMMPS_MACE
 from .mace import MACECalculator
-from .torch_sim import MaceTorchSimModel
+from .mace_torchsim import MaceTorchSimModel
 
 __all__ = [
     "MACECalculator",
